+++ firebase-sessions/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
 
 * [fixed] Handled datastore writes when device has full internal memory more gracefully.
-
+  (GitHub [#5859](https://github.com/firebase/firebase-android-sdk/issues/5859))
+* [fixed] Safely unbind to release service connection.
+  (GitHub [#5869](https://github.com/firebase/firebase-android-sdk/issues/5869))
+* 
 # 1.2.3
 
 * [fixed] Force validation or rotation of FIDs.